requirements.txt
aiogram==2.23.1
python-dotenv>=1.0.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
numpy>=1.24.3
pandas>=2.0.0
python-telegram-bot>=20.3
opencv-python-headless>=4.7.0 
matplotlib